lad fixed..

+++ tellez.aron/I2/lengthlongerdecreasing.cpp
@@ -3,7 +3,7 @@ size_t length_longer_decreasing(const int *arr, const size_t size)
 {
   size_t max = 0;
   size_t len = 1;
-  for (size_t i = 0; i < size - 1; i++)
+  for (size_t i = 0; i < static_cast<unsigned long> (size - 1) ; i++)
   {
     if (arr[i + 1] < arr[i])
     {
